Use NoopLogger by default
