+++ src/index.ts
@@ -1,4 +1,4 @@
 export { Endpoint } from './lib/endpoint';
 export { IbcClient } from './lib/ibcclient';
 export { Link } from './lib/link';
-export { Logger } from './lib/utils';
+export { Logger, NoopLogger } from './lib/logger';
